:recycle: Separate preprocessors
