+++ ice_station_zebra/data_processors/preprocessors/__init__.py
@@ -1,5 +1,6 @@
-from .base import IPreprocessor, NullPreprocessor
 from .icenet_sic import IceNetSICPreprocessor
+from .ipreprocessor import IPreprocessor
+from .null import NullPreprocessor
 
 __all__ = [
     "IPreprocessor",
